Factor LMDB-specifics into lmdb.rs
This manages to remove almost all LMDB-specifics from server.rs
